+++ @TensorSpline/integrateAlongDimension.m
@@ -10,16 +10,9 @@ xiMatrix = reshape(xiPermuted, size(xiPermuted, 1), []);
 
 originalK = K;
 originalTKnot = tKnot;
-transformedSlice = cumsum(BSpline(S=originalK-1, knotPoints=originalTKnot, xi=xiMatrix(:,1), xMean=xMean, xStd=xStd));
-tKnot = transformedSlice.knotPoints;
-K = transformedSlice.K;
-
-transformedMatrix = zeros(numel(transformedSlice.xi), size(xiMatrix, 2), 'like', transformedSlice.xi);
-transformedMatrix(:,1) = transformedSlice.xi;
-for iSlice = 2:size(xiMatrix, 2)
-    transformedSlice = cumsum(BSpline(S=originalK-1, knotPoints=originalTKnot, xi=xiMatrix(:,iSlice), xMean=xMean, xStd=xStd));
-    transformedMatrix(:,iSlice) = transformedSlice.xi;
-end
+[transformedMatrix, tKnot, SIntegrated] = BSpline.integratedSplineState( ...
+    xiMatrix, knotPoints=originalTKnot, S=originalK - 1, xMean=xMean, xStd=xStd);
+K = SIntegrated + 1;
 
 outputSize = size(xiPermuted);
 outputSize(1) = size(transformedMatrix, 1);